Use blosum80 supplied with the nanocorrect
Currently the BLOSUM80 file is expected to be in the current working directory. However, when analysis is done from different directory it's missing and poa fails. The error is gobbled up by DEVNULL. Instead, workout a path of the nanocorrect.py file and use blosum80 file found in that directory.